distribution object for specifying distribution (#285)
Consolidate distribution related options into a single object that can be passed to the distributed_vector constructor. This avoids adding more arguments to the distributed_vector constructor.

Moved halo_bounds into the distribution and added a new one called granularity. Granularity supports implementing a multi-dimensional array with a distributed_vector. For a 2d array, the granularity would be the number of columns, forcing the distribution to be row blocks.